# LABORATORIO/openapi-spec.yaml
openapi: 3.0.3
info:
  title: QA Automation Shop API - Spec Completa
  description: API completa do Northwind Test Platform para importação no Postman (Mock Server)
  version: 2.0.0
  contact:
    name: QA Automation Team
    email: contato@gotasdetecnologia.com.br
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Ambiente Local
  - url: https://northwind-test-platform.vercel.app
    description: Ambiente Produção
...
paths:
...
  /api/v1/orders:
    get:
      tags: [📦 Pedidos]
      summary: Lista pedidos do usuário
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista
          content:
            application/json:
              example:
                data:
                  - id: 1001
                    order_number: ORD-20240906-1001
                    status: delivered
                    total_amount: 599.80
                    created_at: '2024-09-06T11:00:00.000Z'
                    shippers: { company_name: Correios SEDEX }
                pagination: { page: 1, totalPages: 2, total: 11 }
    post:
      tags: [📦 Pedidos]
      summary: Cria novo pedido (checkout)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            example:
              items:
                - product_id: 1
                  quantity: 2
                  unit_price: 299.90
              shipper_id: 1
              address_id: 5
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              example:
                data:
                  id: 1002
                  order_number: ORD-20240906-1002
                  status: pending
                  total_amount: 599.80
                  created_at: '2024-09-06T14:30:00.000Z'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT (sem a palavra "Bearer" na frente)
...